+++ Backend/backend-update-and-delete/backend-update-and-delete_products/components/Product/index.js
@@ -3,12 +3,15 @@ import { useRouter } from "next/router";
 import { ProductCard } from "./Product.styled";
 import { StyledLink } from "../Link/Link.styled";
 import Comments from "../Comments";
+import { useState } from "react";
+import { StyledButton } from "../Button/Button.styled";
 
 export default function Product() {
+  const {isEditMode, setIsEditMode} = useState(false)
   const router = useRouter();
   const { id } = router.query;
 
-  const { data, isLoading } = useSWR(`/api/products/${id}`);
+  const { data, isLoading, mutate } = useSWR(`/api/products/${id}`);
 
   if (isLoading) {
     return <h1>Loading...</h1>;
@@ -16,6 +19,24 @@ export default function Product() {
 
   if (!data) {
     return;
+  }
+
+  async function andleEditProduct(event) {
+    event.preventDefault();
+    const formData = new FormData(event.target);
+    const productData = object.fromEntries(formData);
+    const response = await fetch(`/api/products/${id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(productData),
+    });
+
+    if (response.ok) {
+      mutate();
+    }
+    
   }
 
   return (
@@ -26,6 +47,7 @@ export default function Product() {
         Price: {data.price} {data.currency}
       </p>
       {data.reviews.length > 0 && <Comments reviews={data.reviews} />}
+      <StyledButton type="button" onClick={() => { setIsEditMode(!isEditMode); }}>Edit</StyledButton>
       <StyledLink href="/">Back to all</StyledLink>
     </ProductCard>
   );
