+++ Backend/backend-update-and-delete/backend-update-and-delete_products/pages/[id].js
@@ -1,4 +1,4 @@
-import Product from "../components/product";
+import Product from "../components/Product";
 
 export default function ProductDetailsPage() {
   return (
