+++ Backend/backend-update-and-delete/backend-update-and-delete_products/pages/api/products/[id].js
@@ -14,4 +14,9 @@ export default async function handler(request, response) {
 
     response.status(200).json(product);
   }
+
+  if (request.method === "PUT") {
+    const productToUpdate = await Product.findByIdAndUpdate(id, updatedProduct)
+    response.status(200).json({ status: "Product successfully updated." })
+  }
 }
