+++ Backend/backend-update-and-delete/backend-update-and-delete_products/pages/index.js
@@ -1,6 +1,7 @@
 import ProductList from "../components/ProductList";
 import styled from "styled-components";
 import ProductForm from "../components/ProductForm";
+import useSWR from "swr";
 
 const Heading = styled.h1`
   text-align: center;
@@ -8,6 +9,30 @@ const Heading = styled.h1`
 `;
 
 export default function HomePage() {
+  const { mutate } = useSWR("/api/products");
+
+  async function handleAddProduct(event) {
+    event.preventDefault();
+
+    const formData = new FormData(event.target);
+    const productData = Object.fromEntries(formData);
+
+    const response = await fetch("/api/products", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(productData),
+    });
+
+    if (!response.ok) {
+      console.error(response.status);
+      return;
+    }
+
+    mutate();
+    event.target.reset();
+  }
   return (
     <>
       <Heading>
@@ -16,7 +41,7 @@ export default function HomePage() {
         </span>
         Fish Shop
       </Heading>
-      <ProductForm />
+      <ProductForm onSubmit={handleAddProduct} />
       <hr />
       <ProductList />
     </>
